Move to _model_summary.py, rename to summarize(), auto-detect dtype
